Refatora autenticação para usar LoginForm/RegisterForm unificados
