+++ src/routes.js
@@ -5,9 +5,8 @@ import LogoOnlyLayout from './layouts/LogoOnlyLayout';
 //
 import Blog from './pages/Blog';
 import User from './pages/User';
-import Login from './pages/Login';
+import { LoginForm } from './sections/auth/login';
 import NotFound from './pages/Page404';
-import Register from './pages/Register';
 import Products from './pages/Products';
 import DashboardApp from './pages/DashboardApp';
 import Hortelan360 from './pages/Hortelan360';
@@ -59,7 +58,7 @@ export default function Router() {
       path: 'login',
       element: (
         <RedirectIfAuth>
-          <Login />
+          <LoginForm />
         </RedirectIfAuth>
       ),
     },
@@ -67,7 +66,7 @@ export default function Router() {
       path: 'register',
       element: (
         <RedirectIfAuth>
-          <Register />
+          <LoginForm />
         </RedirectIfAuth>
       ),
     },
